+++ types/grpc/headers.go
@@ -3,4 +3,6 @@ package grpc
 const (
 	// GRPCBlockHeightHeader is the gRPC header for block height.
 	GRPCBlockHeightHeader = "x-cosmos-block-height"
+	// GRPCBlockTimeHeader is the gRPC header for block time.
+	GRPCBlockTimeHeader = "x-cosmos-block-time"
 )